Refactor list items, make thread item a list item with consistent typography
