--- mobile/components/ThreadItem/utils.js
@@ -1,8 +1,0 @@
-// @flow
-export const truncate = (str: string, length: number) => {
-  if (str.length <= length) {
-    return str;
-  }
-  const subString = str.substr(0, length);
-  return subString.substr(0, subString.lastIndexOf(' ')) + '…';
-};
