+++ mobile/views/Community/index.js
@@ -10,7 +10,7 @@ import {
 import getCommunityThreads from '../../../shared/graphql/queries/community/getCommunityThreadConnection';
 import ViewNetworkHandler from '../../components/ViewNetworkHandler';
 import ThreadFeed from '../../components/ThreadFeed';
-import ThreadItem from '../../components/ThreadItem';
+import { ThreadListItem } from '../../components/Lists';
 import { getThreadById } from '../../../shared/graphql/queries/thread/getThread';
 
 import {
@@ -39,7 +39,13 @@ const RemoteThreadItem = compose(getThreadById, withNavigation)(
   ({ data, navigation }) => {
     if (data.loading) return <Text>Loading...</Text>;
     if (!data.thread) return null;
-    return <ThreadItem thread={data.thread} navigation={navigation} />;
+    return (
+      <ThreadListItem
+        activeCommunity={data.thread.community.id}
+        thread={data.thread}
+        navigation={navigation}
+      />
+    );
   }
 );
 
@@ -83,10 +89,16 @@ class Community extends Component<Props> {
                 <ThreadFeedDivider />
 
                 {community.pinnedThreadId && (
-                  <RemoteThreadItem id={community.pinnedThreadId} />
+                  <RemoteThreadItem
+                    id={community.pinnedThreadId}
+                    activeCommunity={community.id}
+                  />
                 )}
                 {community.watercoolerId && (
-                  <RemoteThreadItem id={community.watercoolerId} />
+                  <RemoteThreadItem
+                    id={community.watercoolerId}
+                    activeCommunity={community.id}
+                  />
                 )}
               </Fragment>
             }
